+++ tests/AppBundle/Controller/BlogControllerTest.php
@@ -11,11 +11,11 @@
 
 namespace Tests\AppBundle\Controller;
 
+use AppBundle\DataFixtures\FixturesTrait;
 use AppBundle\Entity\Post;
 use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
 use Symfony\Component\HttpFoundation\Response;
 use Tests\ControllerTestTrait;
-use Tests\FixturesTrait;
 
 /**
  * Functional test for the controllers defined inside BlogController.
